Added Remote Procedure Call example using IDL/ASIO
Updated idl.hpp and mirror_interface.hpp to provide a
better means for getting members on types.  This allows us
to store a rpc_client<Interface> inside a idl::any<Interface> and
therefore code written in terms of idl::any<Interface> has no
idea whether or not the calls go across the network.